Move ProgramManager to its own file.
